Fixes for how format is inferred
The format needs to be inferred when users come to
the document-tasks page from the intent page as
well as when they come from filling in the parts
of an edition.

Because of this we need to ensure the intent is
calculated every time the document-tasks page is
requested.

This also fixes a bug in the code which meant the
set format (either suggested or chosen) was
repeated in the list of alternatives.