Give preview generator its own content cache
This code shared the same data attribute with the
content checker for checking if the content
entered had changed. If it had, the attribute was
set to the latest content.

This meant when the content checker looked at it
it saw no changes and didn't run any checks.

This change means each module uses it's own
attribute so they won't step on each other's toes.