Refactor how format inference is handled
This adds better control over how a user's answer
to our inferring the content format works.

If they agree with it, we should set the format to
the type inferred. If not, they have to choose an
alternative so we should use that but update the
flag for their choice to show the format displayed
was chosen by them.

This and other changes mean we can guarantee the
format will always be set when they leave the
content-settings page. This means the template
logic in the document-tasks page can be made a lot
simpler.