Make parsing collapse raw data into rows for items
In order to display each content item in a single
row we need each field to appear as one value.

This produces a data structure where fields that
are spread across multiple rows in the raw data,
for example pageviews, to be stored as lists of
those values while others that don't change, for
example content_id, will be stored as a single
value.

This enables us to display a final value at render
time by using these different data types in any
transformations needed.